+++ tasks/mpi/voroshilov_v_bivariate_optimization_by_area/perf_tests/main.cpp
@@ -21,7 +21,7 @@ TEST(voroshilov_v_bivariate_optimization_by_area_mpi_perf, test_pipeline_run) {
   std::vector<char> g_vec2(g_str2.length());
   std::copy(g_str2.begin(), g_str2.end(), g_vec2.begin());
   std::vector<std::vector<char>> g_vec({g_vec1, g_vec2});
-  int g_count = g_vec.size();
+  size_t g_count = g_vec.size();
 
   // Search areas:
   std::vector<double> areas_vec({-5.0, 5.0, -5.0, 5.0});
@@ -37,7 +37,7 @@ TEST(voroshilov_v_bivariate_optimization_by_area_mpi_perf, test_pipeline_run) {
   taskDataParallel->inputs_count.emplace_back(1);
   taskDataParallel->inputs.emplace_back(reinterpret_cast<uint8_t *>(&g_count));
   // Constraints-functions:
-  for (int i = 0; i < g_vec.size(); i++) {
+  for (size_t i = 0; i < g_vec.size(); i++) {
     taskDataParallel->inputs_count.emplace_back(g_vec[i].size());
     taskDataParallel->inputs.emplace_back(reinterpret_cast<uint8_t *>(g_vec[i].data()));
   }
@@ -92,7 +92,7 @@ TEST(voroshilov_v_bivariate_optimization_by_area_mpi_perf, test_task_run) {
   std::vector<char> g_vec2(g_str2.length());
   std::copy(g_str2.begin(), g_str2.end(), g_vec2.begin());
   std::vector<std::vector<char>> g_vec({g_vec1, g_vec2});
-  int g_count = g_vec.size();
+  size_t g_count = g_vec.size();
 
   // Search areas:
   std::vector<double> areas_vec({-5.0, 5.0, -5.0, 5.0});
@@ -108,7 +108,7 @@ TEST(voroshilov_v_bivariate_optimization_by_area_mpi_perf, test_task_run) {
   taskDataParallel->inputs_count.emplace_back(1);
   taskDataParallel->inputs.emplace_back(reinterpret_cast<uint8_t *>(&g_count));
   // Constraints-functions:
-  for (int i = 0; i < g_vec.size(); i++) {
+  for (size_t i = 0; i < g_vec.size(); i++) {
     taskDataParallel->inputs_count.emplace_back(g_vec[i].size());
     taskDataParallel->inputs.emplace_back(reinterpret_cast<uint8_t *>(g_vec[i].data()));
   }
